ci(connect): improve release process
