+++ ci/scripts/connect-release-init-npm.js
@@ -4,7 +4,7 @@ const child_process = require('child_process');
 const path = require('path');
 const fs = require('fs');
 
-const { checkPackageDependencies } = require('./npm-dependencies-helper');
+const { checkPackageDependencies, exec, commit, comment } = require('./helpers');
 
 const args = process.argv.slice(2);
 
@@ -17,29 +17,6 @@ if (!allowedSemvers.includes(semver)) {
 }
 
 const ROOT = path.join(__dirname, '..', '..');
-
-const exec = (cmd, params) => {
-    console.log(cmd, ...params);
-
-    const res = child_process.spawnSync(cmd, params, {
-        encoding: 'utf-8',
-        cwd: ROOT,
-    });
-    if (res.status !== 0) {
-        console.log(cmd, ...params);
-        console.log(res);
-    }
-    return res;
-};
-
-const commit = ({ path, message }) => {
-    exec('git', ['add', path]);
-    exec('git', ['commit', '-m', `${message}`]);
-};
-
-const comment = ({ prNumber, body }) => {
-    exec('gh', ['pr', 'comment', `${prNumber}`, '--body', body]);
-};
 
 const getGitCommitByPackageName = (packageName, maxCount = 10) =>
     exec('git', [
